[WIP][c10d][jit] initial torchbind ProcessGroup
ghstack-source-id: 68d8d854bb2a4ff8a9aaa51a1684719b518d5a63
Pull Request resolved: https://github.com/pytorch/pytorch/pull/42944

+++ torch/lib/c10d/ProcessGroup.cpp
@@ -1,8 +1,11 @@
 #include <c10d/ProcessGroup.hpp>
 
 #include <c10/util/Logging.h>
+#include <sys/types.h>
+#include <torch/custom_class.h>
 
 namespace c10d {
+
 
 ProcessGroup::Work::~Work() {}
 
@@ -21,7 +24,7 @@ std::exception_ptr ProcessGroup::Work::exception() const {
   return exception_;
 }
 
-int ProcessGroup::Work::sourceRank() const {
+int64_t ProcessGroup::Work::sourceRank() const {
   throw std::runtime_error(
       "sourceRank() may only be called on work objects "
       "that correspond to a recv or recv-from-any call.");
